changes files tree structure of project
