+++ app/auth/reset-password/page.tsx
@@ -2,7 +2,7 @@ import SetPasswordForm from "@/components/auth/set-password-form"
 
 const NewPassordPage = () => {
   return (
-    <div className="flex justify-center items-center p-10">
+    <div className="flex justify-center min-h-[80vh] items-center p-10">
       <SetPasswordForm />
     </div>
   )
