+++ app/auth/reset/page.tsx
@@ -8,7 +8,7 @@ export const metadata: Metadata = {
 }
 function ResetPasswordPage() {
   return (
-    <div className="flex justify-center items-center p-10">
+    <div className="flex justify-center min-h-[80vh] items-center p-10">
       <ResetForm />
     </div>
   )
