--- components/ui/fonts.ts
@@ -1,11 +1,0 @@
-import { Inter, Cedarville_Cursive, Recursive } from 'next/font/google';
-
-
-export const inter = Inter({ subsets: ['latin'] });
-
-export const recursive = Recursive({ subsets: ['latin']});
-
-export const cedarville = Cedarville_Cursive({
-  weight: '400',
-  subsets: ['latin']
-});
